appstatemock: Use state_dict() to get current state
Saving a reference to the state dictionary at the time of the call to
get_test_interface() will cause the methods of the test interface to use
the wrong object in case the state is updated in a transaction, since
the current state will be in another dictionary object. As such, use
state_dict() to retrive the correct object when needed.