appstate: Add methods set_string() and get_string()
Some state backends might not allow nested transactions. If
set_version() is done in a transaction, the current test code for
transactions (test_transaction()) will not work. Let's prepare to use a
simpler state modification function in transactions. For that, require
AppStateTestInterface to have two very simple methods to update
application state: set_string() for saving a string value and
get_string() for retrieving it back. We will modify test_transaction()
to use those in a future change.